Rewrite metronome.js
Previous version performed badly on mobile device.  This is perhaps
due to inefficient use of the AudioBufferSourceNode.onended handler.
Use of this handler is eliminated.

Unused code retained as a sketch for a fallback has been removed.  It
can be brought back later if there will in fact be a fallback if
Web Audio API is unsupported.

TODO: Subdivisions are scheduled with the beat.  This leads to messy
transitions.  Schedule one note at a time.