Fix subdivision change problem
When subdivisions were active and tempo changed, changes in the
division (duple to triple or vice versa) would not be recognized.

This commit also adds comments and a testing line concerning a
potential problem with beepFunction, namely that the loop is
called twice in immediate succession when start button is clicked
on.