Change only at pattern boundaries
Tempo changes and changes of beat division happen at the upcoming
beat.  For example, in case of a tempo change within a subdivided
rhythm, the last subdivision will be given the full value in the
old tempo; the main beat will arrive where it would in the old
tempo; subsequent divisions occur in the new tempo.

Subdivision button changes text to reflect on/off status.

A number of variable names have been changed for clarity.

lastNoteTime/newNoteTime replace a single variable.  This clarifies
upcoming events becoming past events in scheduling routine.

TODO: long button text causes button to jump a line.  Possibly
"subdivide" should be an independent text item and a button would
toggle between "on" and "off".  "Duple" and "triple" might be
replaced by "2" and "3".