FIX: uneven first beat on mobile
Previously, first note happened right away.  "Right away," however,
could be somewhat delayed.  The second note happened at the right time,
leading to a shortened first beat.

Also, two notes would be scheduled right away, which could pose a
problem with sound/visual integration.

To solve this, the first note is scheduled with a delay of 0.4 sec.
(This value is chosen to be short enough that the user isn't
bothered and long enough to give calculations time to complete.  It
is somewhat arbitrary.)

(Note that there is still an issue with sound/visual integration.
Because beepFunction always schedules future sounds, visual cues will
be off by one.  The function is called before the first sound happens
and before the last sound occurs.)