Add metronome sound, attributions
The metronome sound used is from Audiosoundclips.com.  The mp3
originally consisted of a number of beats.  It has been trimmed
to a single beat.

Attribution has been given in js/metronome.js for this new sound
and for Beep.mp3.

The sound of the metronome now uses only the new percussive sound.