Improve handling of visual count end
Last visual element stays on screen and vanishes. roughly in
tempo.  Starting a new count before the last element vanishes will
not disrupt this by clearing the wrong element.

Note: for testing, the number of on-screen beats has been increased
here to four because problems with beat clearing addressed by this
commit happen when there are more than two beats.