Support multiple sounds
Use promises to load multiple sounds.  The callback for
request.onload has been moved within a function ('init')
called when the last sound is loaded.

Beep.mp3 is used as a placeholder for the off-beat sound.

Right now, two sounds are loaded, but a larger library
will be created.

Other changes:
 - Addition/revision of comments