Improve layout on Safari, use box sizing
Layout on iPhone Safari is now more consistent with other browsers.

This required use of -webkit-appearance: none for buttons, which
is not an ideal solution.  Unfortunately, there seems no other
way to use buttons and remove the protruding rounded edges
which prevent two buttons from sharing a row on the metronome.
(Unless we forego use of the button tag and make our own clickable
elements.)

Metronome options are now presented as an unordered list and texts
are shortened so proper line breaks are assured.

Other changes:
 - space items on side menu further apart so list is more
 usable on mobile