Handle sound load errors gracefully
If one or more sound files fails to load, we substitute
files that did load.  Prefer the default *tock*.

If no file loads, alert the user and do not call init
function.  Metronome will not work, and will not throw
errors if buttons are pressed.

(It might be nice to have some purely display-related
functionality -- show/hide beats, change their number, for
example.  There is no pressing need for this, however.)

Other changes:
 - helpers (hideBeat, showBeat) added for updateDisplay
 - lots of comment revision