Only modify beat makeup from beepFunction
The problem noted before with the beat position marker
going out of range is related to the subdivide event
handler resetting the pattern.

beepFunction checks for the end of a pattern before
processing changes (including to the pattern itself).  If
other functions are allowed to trigger a pattern change,
beepFunction's check can fail.  (Hence the need for the
seemingly extra position reset in resetPattern.)

Other changes:
 - simplify beepFunction slightly
 - rename two variables
 - remove unneeded function (resetPattern) and move its
 functionality to updatePattern.